+++ packages/plugins/vlm-default/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @xpert-ai/plugin-vlm-default
 
+## 8.0.12
+
+### Patch Changes
+
+- Updated dependencies [b8bac1f]
+  - @xpert-ai/plugin-sdk@3.15.12
+
+## 8.0.11
+
+### Patch Changes
+
+- Updated dependencies [aa16ee9]
+  - @xpert-ai/plugin-sdk@3.15.11
+
+## 8.0.10
+
+### Patch Changes
+
+- Updated dependencies [c9d8401]
+  - @xpert-ai/plugin-sdk@3.15.10
+
 ## 8.0.9
 
 ### Patch Changes